requirements.txt
paddleocr==2.10.0
pillow==11.2.1
tqdm==4.67.1
pyyaml==6.02